+++ src/baobab_mtg_products/domain/products/__init__.py
@@ -1,0 +1,21 @@
+"""Modèle métier de référence des produits scellés Magic: The Gathering."""
+
+from baobab_mtg_products.domain.products.commercial_barcode import CommercialBarcode
+from baobab_mtg_products.domain.products.internal_barcode import InternalBarcode
+from baobab_mtg_products.domain.products.internal_product_id import InternalProductId
+from baobab_mtg_products.domain.products.mtg_set_code import MtgSetCode
+from baobab_mtg_products.domain.products.product_instance import ProductInstance
+from baobab_mtg_products.domain.products.product_status import ProductStatus
+from baobab_mtg_products.domain.products.product_type import ProductType
+from baobab_mtg_products.domain.products.serial_number import SerialNumber
+
+__all__ = [
+    "CommercialBarcode",
+    "InternalBarcode",
+    "InternalProductId",
+    "MtgSetCode",
+    "ProductInstance",
+    "ProductStatus",
+    "ProductType",
+    "SerialNumber",
+]